# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "E-commerce backend API for household chemicals, cosmetics, and home goods.",
        "title": "Commerce Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/categories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all categories including inactive (admin)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "List all categories (admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.Category"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new category (admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Create category",
                "parameters": [
                    {
                        "description": "Category data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.CreateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.Category"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/categories/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing category (admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Update category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Category data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.UpdateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.Category"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a category (admin)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Delete category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/subcategories": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new subcategory (admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubCategories"
                ],
                "summary": "Create subcategory",
                "parameters": [
                    {
                        "description": "Subcategory data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.CreateSubCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.SubCategory"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/subcategories/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing subcategory (admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubCategories"
                ],
                "summary": "Update subcategory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Subcategory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Subcategory data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.UpdateSubCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.SubCategory"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a subcategory (admin)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubCategories"
                ],
                "summary": "Delete subcategory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Subcategory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/categories": {
            "get": {
                "description": "Get all active categories with subcategories (public)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "List categories",
                "parameters": [
                    {
                        "type": "string",
                        "default": "ru",
                        "description": "Language (ru or tm)",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.CategoryResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            }
        },
        "/categories/{id}/subcategories": {
            "get": {
                "description": "Get subcategories by category ID (public)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "List subcategories",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "ru",
                        "description": "Language (ru or tm)",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.SubCategoryResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_meransoft_commerce-backend_internal_models.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiError"
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.Pagination"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_meransoft_commerce-backend_internal_models.Category": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "has_subcategories": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name_ru": {
                    "type": "string"
                },
                "name_tm": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_meransoft_commerce-backend_internal_models.CreateCategoryRequest": {
            "type": "object",
            "required": [
                "name_ru",
                "name_tm"
            ],
            "properties": {
                "image_url": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name_ru": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "name_tm": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_meransoft_commerce-backend_internal_models.CreateSubCategoryRequest": {
            "type": "object",
            "required": [
                "name_ru",
                "name_tm",
                "parent_id"
            ],
            "properties": {
                "image_url": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name_ru": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "name_tm": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "parent_id": {
                    "type": "integer"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_meransoft_commerce-backend_internal_models.SubCategory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name_ru": {
                    "type": "string"
                },
                "name_tm": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_meransoft_commerce-backend_internal_models.UpdateCategoryRequest": {
            "type": "object",
            "properties": {
                "image_url": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name_ru": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "name_tm": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_meransoft_commerce-backend_internal_models.UpdateSubCategoryRequest": {
            "type": "object",
            "properties": {
                "image_url": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name_ru": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "name_tm": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}